tie minion registration to kubelet activation
this changes moves minion registration out of the
configure-kubernetes-minion.sh script and puts it in a systemd service
unit that tracks the state of the kubelet.service and
registers/unregisters the minion as appropriate.

If scaling down an autoscaling group permitted an instance to shut down
gracefully this would take care of unregistering minions in response to
down scaling operations, but it doesn't, so this can't.

Change-Id: I6532a9bc08bbb2351a2ae5d29bc884cb3cad3508